requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Computer Vision and Image Processing
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

# Machine Learning / Model Support
mediapipe==0.10.8

# Testing
hypothesis==6.92.1
pytest==7.4.3

# Utilities
python-dateutil==2.8.2

streamlit==1.28.1
Pillow==10.1.0